Develop

+++ bus-office/src/main/java/org/aoju/bus/office/support/excel/RowUtils.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright (c) 2017 aoju.org All rights reserved.
+ * Copyright (c) 2020 aoju.org All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -36,7 +36,7 @@ import java.util.List;
  * Excel中的行{@link Row}封装工具类
  *
  * @author Kimi Liu
- * @version 5.5.0
+ * @version 5.5.1
  * @since JDK 1.8+
  */
 public class RowUtils {
@@ -70,7 +70,7 @@ public class RowUtils {
         if (length < 0) {
             return new ArrayList<>(0);
         }
-        final List<Object> cellValues = new ArrayList<>((int) length);
+        final List<Object> cellValues = new ArrayList<>(length);
         Object cellValue;
         boolean isAllNull = true;
         for (short i = 0; i < length; i++) {
